move the DiagnosticManager initialization from Tracer to Instrumentation
