+++ src/Datadog.Trace/Tracer.cs
@@ -384,36 +384,6 @@ namespace Datadog.Trace
             return _agentWriter.FlushTracesAsync();
         }
 
-#if NETSTANDARD
-        internal void StartDiagnosticObservers()
-        {
-            // instead of adding a hard dependency on DiagnosticSource,
-            // check if it is available before trying to use it
-            var type = Type.GetType("System.Diagnostics.DiagnosticSource, System.Diagnostics.DiagnosticSource", throwOnError: false);
-
-            if (type == null)
-            {
-                Log.Warning("DiagnosticSource type could not be loaded. Disabling diagnostic observers.");
-            }
-            else
-            {
-                // don't call this method unless the necessary types are available
-                StartDiagnosticObserversInternal();
-            }
-        }
-
-        internal void StartDiagnosticObserversInternal()
-        {
-            DiagnosticManager?.Stop();
-
-            Log.Debug("Starting DiagnosticManager");
-
-            var observers = new List<DiagnosticObserver> { new AspNetCoreDiagnosticObserver() };
-            var diagnosticManager = new DiagnosticManager(observers);
-            diagnosticManager.Start();
-        }
-#endif
-
         internal async Task WriteDiagnosticLog()
         {
             string agentError = null;
